[alloc @ alla@cyber.com.au-20071031160923-rzi0o958i3orfj18] Did a lot of work on improving the soap services stuff. Added external soap hooks for add_timeSheetItem_by_task() and get_task_comments().

+++ shared/lib/class_session.inc.php
@@ -57,11 +57,13 @@ class Session {
 
 
   // Call this in a login page to start session 
-  function Start($row) {
+  function Start($row,$nuke_prev_sessions=true) {
     $this->key = md5($row["personID"]."mix it up#@!".md5(mktime().md5(microtime())));
     $this->Put("key2", $this->key2);
     $this->Put("session_started", mktime());
-    $this->db->query("DELETE FROM sess WHERE personID = %d",$row["personID"]);
+    if ($nuke_prev_sessions) {
+      $this->db->query("DELETE FROM sess WHERE personID = %d",$row["personID"]);
+    }
     $this->db->query("INSERT INTO sess (sessID,sessData,personID) VALUES (%s,%s,%d)"
                              ,$this->key, $this->Encode($this->session_data), $row["personID"]);
     $this->Put("username" ,strtolower($row["username"]));
